[MOV] tools: move phone_validation.py into odoo.tools
This commit has for main purpose to make available the phone_validation
tools for every module without needed to base on phone_validation odoo
module.
We do that so website can implement phone formatting based on already
known dependencies we have in backend.

task-3700790
